streamlit_app/requirements.txt
streamlit
pandas
biopython
matplotlib
seaborn
plotly
